ShallowCopy on NetworkMessage
In the event that a NetworkMessage must be sent to various target peers
but also must be encrypted uniquely then we can ShallowCopy the message
to preform encryption on each. This saves the step of serializeing N
times for each peer.

In the event that encryption is not required no copying is needed. A
message can be serialized once and send to N clients.