Fixed Fault from exception bringing down network thread
Exceptions were causing the internal network thread to stop reading
messages. This is OK for clients but not for servers.

New functionaly spoofs a Status Message that ends up on the client that
is blamed for the exception. Exceptions still get pushed to the
OnException handler for the lidgren client, so it will log too. No
exception information is provided to the peer.