Added BufferWaitUntil
This is a decorator that will allow you to buffer until a certian size
threshold is reached. For example if you have a header that is written
but also need blocksize crypto to work with the payload you may need to
buffer until the full payload can be written at once.