Added IBytesWrittable to IManagedNetworkClient (Hack)
Ok, so this is a hack but a much needed "feature". Sometimes, in very
rare cases, it is important to bypass the high level networking layer
that is strongly typed and write bytes to it directly. This could be for
performance reasons or for reasons of compatibility. For example WoW has
a case where nearly 20 packets are the same except for an opcode.
Instead of creating 20 handlers it may be easier to handle this at a
lower level.